dando erro em queryResult1()

+++ src/main/java/com/museu/model/Emprestados.java
@@ -15,7 +15,7 @@ public class Emprestados {
     @Id
     @NotNull
     @Column(name = "numobj4")
-    private Long numObj4;
+    private int numObj4;
 
     @NotNull
     @Column(name = "dataemprestimo")
@@ -31,18 +31,18 @@ public class Emprestados {
 
     public Emprestados(){}
 
-    public Emprestados(Long numObj4, Timestamp dataEmprestimo, Timestamp dataDevolucao, String nomeColPert) {
+    public Emprestados(int numObj4, Timestamp dataEmprestimo, Timestamp dataDevolucao, String nomeColPert) {
         this.numObj4 = numObj4;
         this.dataEmprestimo = dataEmprestimo;
         this.dataDevolucao = dataDevolucao;
         this.nomecolpert = nomecolpert;
     }
 
-    public Long getNumObj4() {
+    public int getNumObj4() {
         return numObj4;
     }
 
-    public void setNumObj4(Long numObj4) {
+    public void setNumObj4(int numObj4) {
         this.numObj4 = numObj4;
     }
