+++ src/main/java/com/museu/model/Outros.java
@@ -11,7 +11,7 @@ public class Outros {
     @Id
     @NotNull
     @Column(name = "numobj3")
-    private Long numObj3;
+    private int numObj3;
 
     @NotNull
     @Column(name = "tipo")
@@ -19,16 +19,16 @@ public class Outros {
 
     public Outros(){}
 
-    public Outros(Long numObj3, String tipo) {
+    public Outros(int numObj3, String tipo) {
         this.numObj3 = numObj3;
         this.tipo = tipo;
     }
 
-    public Long getNumObj3() {
+    public int getNumObj3() {
         return numObj3;
     }
 
-    public void setNumObj3(Long numObj3) {
+    public void setNumObj3(int numObj3) {
         this.numObj3 = numObj3;
     }
